Rename PR comment action
